requirements.txt

python-dotenv>=1.0.1
requests>=2.31.0
openai>=1.6.1
